LPS-57070 Move
