+++ modules/apps/search/search-facets/src/com/liferay/search/facets/util/SearchFacetTracker.java
@@ -14,7 +14,7 @@
 
 package com.liferay.search.facets.util;
 
-import com.liferay.search.facets.facet.SearchFacet;
+import com.liferay.search.facets.SearchFacet;
 
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
