+++ modules/apps/search/search-web/src/META-INF/resources/init.jsp
@@ -89,7 +89,7 @@ page import="com.liferay.portlet.asset.model.AssetRendererFactory" %><%@
 page import="com.liferay.portlet.asset.service.AssetCategoryLocalServiceUtil" %><%@
 page import="com.liferay.portlet.asset.service.AssetEntryLocalServiceUtil" %><%@
 page import="com.liferay.portlet.asset.service.permission.AssetCategoryPermission" %><%@
-page import="com.liferay.search.facets.facet.SearchFacet" %><%@
+page import="com.liferay.search.facets.SearchFacet" %><%@
 page import="com.liferay.search.facets.util.SearchFacetTracker" %><%@
 page import="com.liferay.search.web.constants.SearchPortletKeys" %><%@
 page import="com.liferay.search.web.context.SearchDisplayContext" %><%@
