+++ modules/apps/search/search-web/src/com/liferay/search/web/context/SearchDisplayContext.java
@@ -24,7 +24,7 @@ import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PropsValues;
-import com.liferay.search.facets.facet.SearchFacet;
+import com.liferay.search.facets.SearchFacet;
 import com.liferay.search.facets.util.SearchFacetTracker;
 
 import java.util.List;
