+++ modules/apps/search/search-web/src/com/liferay/search/web/facet/FolderSearchFacet.java
@@ -20,8 +20,8 @@ import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.facet.MultiValueFacet;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.search.facets.facet.BaseJSPSearchFacet;
-import com.liferay.search.facets.facet.SearchFacet;
+import com.liferay.search.facets.BaseJSPSearchFacet;
+import com.liferay.search.facets.SearchFacet;
 
 import javax.portlet.ActionRequest;
